Address code review issues: bugs, cleanup, type safety, and tests
- Add DOLT_DATABASES setting, @login_required, and db_alias validation in demo
- Fix OrderItem.subtotal to return Decimal instead of float
- Remove single-DB commits/pull views and templates from DoltAdminMixin
- Consolidate _conn/_get_connection: models.py imports from services.py
- Add logging to silent exception blocks in admin.py
- Fix pluralization logic with explicit name map
- Move dolt_autocommit/get_author_from_request to direct imports in __init__.py
- Add return type annotation to release.py main()
- Remove duplicate dolt_databases fixture from conftest.py
- Fix all mypy errors across source and test files
- Add tests for response_add/response_change, register_default_dolt_admin,
  get_app_list regrouping, and register_branch_extension

+++ demo/demo_app/models.py
@@ -5,6 +5,8 @@ This module defines models for two separate Dolt databases:
 - Order models (orders): Customers and orders
 """
 
+from decimal import Decimal
+
 from django.db import models
 
 
@@ -146,6 +148,6 @@ class OrderItem(models.Model):
         return f"{self.quantity}x {self.product_name}"
 
     @property
-    def subtotal(self) -> float:
+    def subtotal(self) -> Decimal:
         """Calculate subtotal for this line item."""
-        return float(self.quantity * self.unit_price)
+        return self.quantity * self.unit_price
